+++ src/apps/kfone-selfcare/src/api/index.js
@@ -17,6 +17,5 @@
  */
 
 export * from './account-verification';
-export * from './auth';
 export * from './usage-data';
 export * from './user-interactions';
