+++ src/apps/kfone-selfcare/src/components/index.js
@@ -18,11 +18,6 @@
 
 export * from './auth';
 export * from './buttons';
-export * from './cards';
-export * from './feedback';
-export * from './forms';
 export * from './loaders';
-export * from './navigation';
-export * from './sections';
 export * from './carousel';
 export * from './error-boundary';
