+++ src/apps/kfone-selfcare/src/components/loaders/index.js
@@ -17,5 +17,4 @@
  */
 
 export * from './loading';
-export * from './spinner';
 export * from './section-loader';
