+++ src/apps/kfone-selfcare/src/layouts/index.js
@@ -19,4 +19,3 @@
 export * from './auth-layout';
 export * from './customer-portal-layout';
 export * from './error-layout';
-export * from './general-layout';
