--- src/apps/kfone-selfcare/src/pages/customer-portal/explore-page/explore-page.jsx
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import React from 'react';
-import {CustomerPortalLayout} from '../../../layouts';
-import {Carousel} from '../../../components';
-import {IOTDevicesSection, MobilePlansSection, PhonesSection, TVPlansSection} from './sections';
-
-export const ExplorePage = () => {
-  return (
-    <CustomerPortalLayout>
-      <Carousel />
-      <MobilePlansSection />
-      <PhonesSection />
-      <TVPlansSection />
-      <IOTDevicesSection />
-    </CustomerPortalLayout>
-  );
-};
