+++ src/apps/kfone-selfcare/src/pages/explore-page/explore-page.jsx
@@ -16,11 +16,19 @@
  * under the License.
  */
 
-export const validateEmail = email => {
-  // eslint-disable-next-line no-useless-escape
-  if (/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email)) {
-    return true;
-  } else {
-    return false;
-  }
+import React from 'react';
+import {CustomerPortalLayout} from '../../layouts';
+import {Carousel} from '../../components';
+import {IOTDevicesSection, MobilePlansSection, PhonesSection, TVPlansSection} from './sections';
+
+export const ExplorePage = () => {
+  return (
+    <CustomerPortalLayout>
+      <Carousel />
+      <MobilePlansSection />
+      <PhonesSection />
+      <TVPlansSection />
+      <IOTDevicesSection />
+    </CustomerPortalLayout>
+  );
 };
