+++ src/apps/kfone-selfcare/src/pages/explore-page/sections/mobile-plans-section.jsx
@@ -20,7 +20,7 @@
 
 import {useRef} from 'react';
 import {useAuthContext} from '@asgardeo/auth-react';
-import {recordUserInteractions} from '../../../../api';
+import {recordUserInteractions} from '../../../api';
 import {MdCheck} from 'react-icons/md';
 
 export const MobilePlansSection = () => {
