+++ src/apps/kfone-selfcare/src/utils/index.js
@@ -17,5 +17,3 @@
  */
 
 export * from './date-time-utils';
-export * from './dom-utils';
-export * from './validation-utils';
