+++ src/apps/kfone-website/src/components/loaders/index.js
@@ -16,6 +16,4 @@
  * under the License.
  */
 
-export * from './loading';
 export * from './spinner';
-export * from './section-loader';
